+++ selfstudy-framework/src/main/java/com/lxy/framework/security/util/UserIdUtil.java
@@ -5,7 +5,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-
 /**
  * 获取用户id工具类
  *
@@ -15,15 +14,15 @@ import org.springframework.security.core.context.SecurityContextHolder;
  */
 public class UserIdUtil {
 
-    public static int getUserId() {
-        int userId = -1;
-        //获取SecurityContextHolder中的用户id
+    public static long getUserId() {
+        long userId = -1L;
+        // 获取SecurityContextHolder中的用户id
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication == null) {
             return userId;
         }
         if (authentication instanceof UsernamePasswordAuthenticationToken authenticationToken) {
-            StatelessUser principal = (StatelessUser) authenticationToken.getPrincipal();
+            StatelessUser principal = (StatelessUser)authenticationToken.getPrincipal();
             userId = principal.getUserId();
 
         }
